Extract workspace MCP APIs
