+++ Tests/QuillCodeParityTests/ParityMCPGateTests.swift
@@ -4,6 +4,7 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesMCPSupportTypes() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let mcpSurfaceText = try Self.appSourceText(named: "QuillCodeMCPSurface.swift")
+        let mcpModelText = try Self.appSourceText(named: "WorkspaceModelMCP.swift")
         let mcpRequestText = try Self.appSourceText(named: "WorkspaceMCPRequests.swift")
         let mcpRuntimeText = try Self.appSourceText(named: "WorkspaceMCPRuntime.swift")
         let mcpLauncherText = try Self.appSourceText(named: "WorkspaceMCPServerLauncher.swift")
@@ -27,6 +28,13 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(mcpCatalogText.contains("struct WorkspaceMCPToolCatalog"), "MCP dynamic tool descriptions should live in a focused catalog file.")
         XCTAssertTrue(mcpRuntimeText.contains("WorkspaceMCPToolCatalog("), "MCP runtime should delegate dynamic tool definitions to the catalog.")
         XCTAssertTrue(mcpRuntimeText.contains("static func executionOverride"), "MCP dynamic tool routing should live in the runtime.")
+        XCTAssertTrue(mcpModelText.contains("extension QuillCodeWorkspaceModel"), "MCP workspace APIs should live in a focused model extension.")
+        XCTAssertTrue(mcpModelText.contains("func startMCPServer"), "MCP start orchestration should live in the focused model extension.")
+        XCTAssertTrue(mcpModelText.contains("func stopMCPServer"), "MCP stop orchestration should live in the focused model extension.")
+        XCTAssertTrue(mcpModelText.contains("applyMCPRuntimeResult"), "MCP result application should be named and shared by start/stop.")
+        XCTAssertFalse(modelText.contains("func startMCPServer"), "WorkspaceModel.swift should not own MCP start APIs.")
+        XCTAssertFalse(modelText.contains("func stopMCPServer"), "WorkspaceModel.swift should not own MCP stop APIs.")
+        XCTAssertFalse(modelText.contains("finishMCPServerProcess"), "WorkspaceModel.swift should not own MCP termination callbacks.")
         XCTAssertFalse(modelText.contains("public struct ExtensionsState"), "WorkspaceModel should not own MCP extension state.")
         XCTAssertFalse(modelText.contains("public enum MCPServerLifecycleStatus"), "WorkspaceModel should not own MCP lifecycle state.")
         XCTAssertFalse(modelText.contains("public struct MCPServerProbeSummary"), "WorkspaceModel should not own MCP probe summaries.")
